refactor/move createSfWebRequestSubsetMock to Trait
