+++ composer.json
@@ -1,8 +1,8 @@
 {
-    "name": "brnc/psr7-symfony1-adapter",
+    "name":        "brnc/psr7-symfony1-adapter",
     "description": "Partial PSR-7 Adapters for Symfony 1.5",
-    "keywords": ["symfony1", "http-message", "http", "message", "psr-7"],
-    "license": "MIT",
+    "keywords":   ["symfony1", "http-message", "http", "message", "psr-7"],
+    "license":     "MIT",
     "authors": [
         {
             "name": "ebln",
@@ -15,7 +15,7 @@
         }
     },
     "require": {
-        "php": "^5.5 || ^7.0",
+        "php":                      "^5.5 || ^7.0",
         "brnc/http-message-subset": "^0.0.4"
     },
     "archive": {
